+++ src/main/java/crud_app/dao/UserDAO.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface UserDAO {
     void addUser(User user);
     List<User> getAllUsers();
-    void deleteUser(int id);
-    User findUserById(int id);
+    void deleteUser(long id);
+    User findUserById(long id);
 }
