+++ src/main/java/crud_app/service/UserService.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface UserService {
     void addUser(User user);
     List<User> getAllUsers();
-    void deleteUser(int id);
-    User findUserById(int id);
+    void deleteUser(long id);
+    User findUserById(long id);
 }
